SH/Git: Clean up. Add git commands. Work better on Mac.

+++ install.sh
@@ -2,7 +2,7 @@
 #
 # Installation script creates links in the user's home directory
 
-FILES=("vimrc" "vim" "tmux" "tmux.conf" "inputrc" "shrc.d" "gitconfig")
+FILES=("vimrc" "vim" "tmux" "tmux.conf" "inputrc" "shrc.d" "gitconfig.common")
 cat <<EOF
 This script will create soft links to the following files in the user's home
 directory:
@@ -30,10 +30,11 @@ SCRIPTPATH=`pwd`
 popd > /dev/null
 
 function create_link {
-    local DEST="$HOME/.$1"
-    local SRC="$SCRIPTPATH/$1"
+    local SRC="$1"
+    local DEST="$2"
     [[ -e "$DEST" && ! -L "$DEST" ]] && { echo "ERROR: $DEST already exists. Skipping."; return 1; }
-    echo "Linking $DEST -> $SRC"
+    [[ -L "$DEST" && $(readlink "$DEST") -ef "$SRC" ]] && { echo "SKIPPING: $DEST already points to correct file."; return 0; }
+    echo "INFO: Linking $DEST -> $SRC"
     if [[ $(uname -s) == "Darwin" ]]; then
         ln -shf "$SRC" "$DEST"
     else
@@ -42,8 +43,11 @@ function create_link {
 }
 
 function append_shrc() {
-    echo "Updating .$1 to include .shrc.d processing"
+    echo "INFO: Updating .$1 to include .shrc.d processing"
+
+    echo "# SM: -- Begin offload" >> $HOME/.$1
     cat "$SCRIPTPATH/$1" >> $HOME/.$1
+    echo "# SM: -- End offload" >> $HOME/.$1
 }
 
 function check_shrc() {
@@ -52,21 +56,47 @@ function check_shrc() {
         if ! grep 'SM: -- Begin offload' $HOME/.$1 > /dev/null ; then
             append_shrc "$1"
         else
-            echo "Replacing .$1 to update .shrc.d processing. Backup at .$1.old"
-            sed -i.old '/SM: -- Begin offload/,/SM: -- End offload/ {//!d}; /SM: -- Begin offload/r'$1 $HOME/.$1
+            echo "INFO: Replacing .$1 to update .shrc.d processing. Backup at .$1.old"
+            sed -i.old '/SM: -- Begin offload/,/SM: -- End offload/ {//!d;}; /SM: -- Begin offload/r'$1 $HOME/.$1
         fi
     else
         append_shrc "$1"
     fi
 }
 
+echo -e "\n## Offloading zshrc/bashrc\n"
+
 check_shrc zshrc
 check_shrc bashrc
 
+echo -e "\n## Linking dotfiles\n"
+
 # Create softlinks
 for f in "${FILES[@]}"; do
-    create_link "$f"
+    create_link "$SCRIPTPATH/$f" "$HOME/$f"
 done
+
+echo -e "\n## Adding git common commands\n"
+
+# Setup git
+if git config --get-all include.path|grep -q "$HOME/.gitconfig.common"; then
+    echo "SKIPPING: Git already set up"
+else
+    git config --global --add include.path "$HOME/.gitconfig.common"
+    echo "INFO: Added ~/.gitconfig.common to git global include path"
+fi
+
+echo -e "\n## Linking bin files\n"
+
+# Setup bin files
+for f in "$SCRIPTPATH/bin/"*; do
+    create_link "$f" "$HOME/bin/$(basename $f)"
+done
+
+echo -e "\n## Linking man files\n"
+
+# Setup man files
+create_link "$SCRIPTPATH/man" "$HOME/man"
 
 # Offer to install VIM bundles
 echo
